requirements.txt
Pillow
python-barcode
requests
